+++ PriFiles/Filters/AlphaBetaLeastSquaresFilter.cpp
@@ -12,8 +12,8 @@ void AlphaBetaLeastSquaresFilter::initialization(QVector<Target> array) {
 
     m_numberOfSteps = array.size();
 
-    m_alpha = 2.0 * ((2.0 * m_numberOfSteps) - 1.0) / (m_numberOfSteps * (m_numberOfSteps + 1.0));
-    m_beta = 6.0 / (m_numberOfSteps * (m_numberOfSteps + 1.0));
+    m_alpha = static_cast<qreal>(2.0 * ((2.0 * m_numberOfSteps) - 1.0) / (m_numberOfSteps * (m_numberOfSteps + 1.0)));
+    m_beta = static_cast<qreal>(6.0 / (m_numberOfSteps * (m_numberOfSteps + 1.0)));
 
     for(int i = 0; i < array.size(); i++) {
         if(m_targets.size() < m_numberOfTargetsToExtrapolation) {
@@ -28,8 +28,8 @@ Target AlphaBetaLeastSquaresFilter::filterMeasuredValue(Target measurement) {
     Target extrapolation = extrapolateOnTime(measurement.time);
 
     if(m_numberOfSteps < m_maximumNumberOfSteps) {
-        m_alpha = 2.0 * ((2.0 * m_numberOfSteps) - 1.0) / (m_numberOfSteps * (m_numberOfSteps + 1.0));
-        m_beta = 6.0 / (m_numberOfSteps * (m_numberOfSteps + 1.0));
+        m_alpha = static_cast<qreal>(2.0 * ((2.0 * m_numberOfSteps) - 1.0) / (m_numberOfSteps * (m_numberOfSteps + 1.0)));
+        m_beta = static_cast<qreal>(6.0 / (m_numberOfSteps * (m_numberOfSteps + 1.0)));
     }
 
     QPointF filteredCoordinate = extrapolation.coordinate +
